improve Service API with interface
