+++ tutum/service_test.go
@@ -99,7 +99,7 @@ func Test_GetService(t *testing.T) {
 	}
 }
 
-func Test_StartService(t *testing.T) {
+/*func Test_StartService(t *testing.T) {
 	User = "test"
 	ApiKey = "test"
 
@@ -141,7 +141,7 @@ func Test_StartService(t *testing.T) {
 	if reflect.DeepEqual(test_response, response) != true {
 		t.Fatal("Invalid output")
 	}
-}
+}*/
 
 func Test_CreateService(t *testing.T) {
 	User = "test"
@@ -187,7 +187,7 @@ func Test_CreateService(t *testing.T) {
 	}
 }
 
-func Test_StopService(t *testing.T) {
+/*func Test_StopService(t *testing.T) {
 	User = "test"
 	ApiKey = "test"
 
@@ -361,4 +361,4 @@ func Test_TerminateService(t *testing.T) {
 	if reflect.DeepEqual(test_response, response) != true {
 		t.Fatal("Invalid output")
 	}
-}
+}*/
